requirements.txt
svgwrite==1.4.3
pillow==10.4.0
numpy==2.0.1
scipy==1.14.1
streamlit==1.37.1
